+++ components/color-picker/__tests__/index.test.tsx
@@ -3,7 +3,7 @@ import { mount, render } from 'enzyme'
 import ColorPicker from '../index'
 import ColorPickerPanel from '../color-picker-panel'
 import ConfigProvider from '../../config-provider/index'
-import { BorderTypes, IColorPickerProps } from '../interface'
+import { BorderTypes, IColorPickerInputRef, IColorPickerProps } from '../interface'
 import mountTest from '../../../tests/shared/mountTest'
 import { act } from 'react-dom/test-utils'
 import { sleep } from '../../../tests/utils'
@@ -117,12 +117,8 @@ describe('ColorPicker', () => {
     ).toBeFalsy()
 
     // className
-    expect(
-      underlineWrapper.find('.kd-color-picker-container .kd-color-picker-input').at(2).hasClass('my-color-picker'),
-    ).toBeTruthy()
-    expect(
-      borderedWrapper.find('.kd-color-picker-container .kd-color-picker-input').at(2).hasClass('my-color-picker'),
-    ).toBeTruthy()
+    expect(underlineWrapper.find('.kd-color-picker-container').hasClass('my-color-picker')).toBeTruthy()
+    expect(borderedWrapper.find('.kd-color-picker-container').hasClass('my-color-picker')).toBeTruthy()
 
     // placeholder
     expect(underlineWrapper.find('.kd-input-underline').prop('placeholder')).toEqual('#')
@@ -138,8 +134,8 @@ describe('ColorPicker', () => {
     expect(underlineWrapper.find('.kd-input-wrapper-underline').prop('style')).toBeFalsy()
     underlineWrapper.setProps({ style: { height: '40px' } })
     underlineWrapper.update()
-    expect(underlineWrapper.find('.kd-input-wrapper-underline').prop('style')).toEqual({ height: '40px' })
-    expect(underlineWrapper.find('.kd-input-wrapper-underline')).toHaveStyle('height', '40px')
+    expect(underlineWrapper.find('.kd-color-picker-container').prop('style')).toEqual({ height: '40px' })
+    expect(underlineWrapper.find('.kd-color-picker-container')).toHaveStyle('height', '40px')
 
     // default prefixIcon
     expect(underlineWrapper.find('.kd-input-prefix').find('.kd-color-picker-icon')).toHaveClassName(
@@ -243,6 +239,18 @@ describe('ColorPicker', () => {
       '100%',
     )
 
+    // disabled
+    underlineDefValueWrapper.setProps({ disabled: true })
+    borderedDefValueWrapper.setProps({ disabled: true })
+    underlineDefValueWrapper.update()
+    borderedDefValueWrapper.update()
+    expect(underlineDefValueWrapper.find('.kd-input-wrapper-disabled')).toExist()
+    expect(borderedDefValueWrapper.find('.kd-input-wrapper-disabled')).toExist()
+    underlineDefValueWrapper.setProps({ disabled: false })
+    borderedDefValueWrapper.setProps({ disabled: false })
+    underlineDefValueWrapper.update()
+    borderedDefValueWrapper.update()
+
     // value
     underlineDefValueWrapper.setProps({ value: 'red' })
     borderedDefValueWrapper.setProps({ value: 'blue' })
@@ -344,6 +352,11 @@ describe('ColorPicker', () => {
     defaultOpenWrapper.update()
     expect(defaultOpenWrapper.find('.kd-color-picker-panel-historical-color-box')).not.toExist()
     expect(defaultOpenWrapper.find('.kd-color-picker-panel-color-box-title')).not.toExist()
+
+    // popperClassName
+    defaultOpenWrapper.setProps({ popperClassName: 'color-picker-test' })
+    defaultOpenWrapper.update()
+    expect(defaultOpenWrapper.find('.color-picker-test')).toExist()
 
     // showColorTransfer
     defaultOpenWrapper.setProps({ showColorTransfer: false })
@@ -675,11 +688,10 @@ describe('ColorPicker', () => {
     expect(wrapper.find('.kd-color-picker-panel-switch').children().at(0).text()).toEqual('followFunctionalColor')
   })
 
-  //! ref api暂未开发，antd、TDesign未提供
   // 11.ref test
-  // it('should get correct dom from ref of props', () => {
-  //   const ref = React.createRef()
-  //   mount(<ColorPicker ref={ref}></ColorPicker>)
-  //   expect(ref.current instanceof HTMLInputElement).toBe(true)
-  // })
+  it('should get correct dom from ref of props', () => {
+    const ref = React.createRef<IColorPickerInputRef>()
+    mount(<ColorPicker ref={ref}></ColorPicker>)
+    expect(ref?.current?.dom instanceof HTMLDivElement).toBe(true)
+  })
 })